Move rds client to new package
